added cornerpin.py
cornerpin.py is a tool to apply the Adobe Premiere Pro Corner Pin
Effekt, but in a faster Way.

It gets the coordinates of the corners as argument, computes linear
functions between the corners and generates opencv.remap matrixes
to finally apply opencv remap.

By adding other function-generators, most of the code can be
re-used when implementing the straighten program or even more
complex mappings.

Currently, the Cornerpin result here differs a bit (too much) from
Premieres result. I am not sure weather it is my fault or premiere's.