reworked cornerpin.py, updated README
Now opencv nativ homographie functions are used to generate the x- and
ymap for the remap functions. This is faster than using cv2.warpPerspective
the whole time.

The old, self build, version was wrong. I reworked the code a bit and
add it as cornerpin_own_dontwork.py to maybe sometimes understand it
better and be able to modify its code to let it work the right way.

Also, there is now the image cpcal.png, a test image in 1920x1080.